break canonicalizer into submodules to make it easier to comprehend
